encryptionUtils.js
Code provides utility functions for encryption and decryption using the AES-256-CBC algorithm in Node.js. It includes the following functions:

generateRandomNumber: This function generates a secure random number using the crypto.randomBytes method and applies a SHA256 hash to it using crypto-js to ensure randomness and uniqueness.

encrypt: This function takes a plain text string as input and generates an encrypted version of the text using AES-256-CBC encryption. It generates a secure random salt and initialization vector (IV) using generateRandomNumber and crypto.randomBytes, respectively. It then uses the crypto.createCipheriv function to create a cipher object with the specified algorithm, salt, and IV. The plain text is encrypted using the cipher.update method, and the final encrypted data is obtained by calling cipher.final. The function returns the encrypted text, salt, and IV.

decrypt: This function takes an encrypted text string, salt, and IV as input and decrypts the text using AES-256-CBC decryption. It creates a decipher object using crypto.createDecipheriv with the specified algorithm, salt, and IV. The encrypted text is decrypted using the decipher.update method, and the final decrypted data is obtained by calling decipher.final. The function returns the decrypted plain text.

It's important to note that the code assumes the use of Node.js and requires the crypto module and crypto-js library to be installed. Additionally, encryption and decryption rely on secure random number generation and the appropriate management of encryption keys and other cryptographic practices to ensure the security of the encrypted data.